combine new validation tests into validate.py
