+++ projects/rocprofiler-sdk/tests/rocprofv3/rocpd/CMakeLists.txt
@@ -322,9 +322,14 @@ rocprofiler_add_integration_validate_test(
          ${CMAKE_CURRENT_BINARY_DIR}/rocpd-output-data/out_kernel_trace.csv
          ${CMAKE_CURRENT_BINARY_DIR}/rocpd-output-data/out_memory_allocation_trace.csv
          ${CMAKE_CURRENT_BINARY_DIR}/rocpd-output-data/out_regions_trace.csv
+         --csv-input-summary
+         ${CMAKE_CURRENT_BINARY_DIR}/rocpd-output-data/summary/out_kernels_summary.csv
+         --csv-input-truncated
+         ${CMAKE_CURRENT_BINARY_DIR}/rocpd-output-data/summary/out_trunc_kernels_summary.csv
     TIMEOUT 120
     LABELS "integration-tests;rocpd"
-    FIXTURES_REQUIRED rocprofv3-test-rocpd-generation)
+    FIXTURES_REQUIRED rocprofv3-test-rocpd-generation
+                      rocprofv3-test-rocpd-generation-truncated)
 
 # Validate region category filtering
 rocprofiler_add_integration_validate_test(
@@ -348,23 +353,6 @@ rocprofiler_add_integration_validate_test(
     TIMEOUT 60
     LABELS "integration-tests;rocpd;region-category"
     FIXTURES_REQUIRED rocprofv3-test-rocpd-region-category)
-
-rocprofiler_add_integration_validate_test(
-    rocprofv3-test-rocpd-truncate-kernels
-    TEST_PATHS validate_truncate_kernels.py
-    COPY conftest.py
-    CONFIG pytest.ini
-    ARGS --json-input
-         ${CMAKE_CURRENT_BINARY_DIR}/rocpd-input-data/out_results.json
-         --csv-input-truncated
-         ${CMAKE_CURRENT_BINARY_DIR}/rocpd-output-data/summary/out_trunc_kernels_summary.csv
-         --csv-input
-         ${CMAKE_CURRENT_BINARY_DIR}/rocpd-output-data/summary/kernels_summary.csv
-    TIMEOUT 120
-    LABELS "integration-tests;rocpd"
-    FIXTURES_REQUIRED
-        rocprofv3-test-rocpd-generation-truncated
-        rocprofv3-test-rocpd-generation)
 
 #########################################################################################
 #
