+++ projects/rocprofiler-sdk/tests/rocprofv3/rocpd/conftest.py
@@ -92,6 +92,11 @@ def pytest_addoption(parser):
         action="store",
         help="Path to truncated kernels summary CSV file.",
     )
+    parser.addoption(
+        "--csv-input-summary",
+        action="store",
+        help="Path to full kernels summary CSV file.",
+    )
 
     pd.set_option("display.width", 2000)
     # increase debug display of pandas dataframes
@@ -227,21 +232,11 @@ def csv_kernels_truncated(request):
 
 @pytest.fixture
 def csv_kernels_full(request):
-    """Load full kernels summary CSV file from existing --csv-input list"""
-    filenames = request.config.getoption("--csv-input")
-    if not filenames:
-        pytest.skip("--csv-input not provided")
+    """Load full kernels summary CSV file from --csv-input-summary option"""
+    filename = request.config.getoption("--csv-input-summary")
+    if not filename:
+        pytest.skip("--csv-input-summary not provided")
+    if not os.path.exists(filename):
+        raise FileExistsError(f"{filename} does not exist")
 
-    # Find kernels_summary.csv in the list
-    kernel_summary_file = None
-    for f in filenames:
-        if "kernels_summary.csv" in f:
-            kernel_summary_file = f
-            break
-
-    if not kernel_summary_file:
-        pytest.skip("kernels_summary.csv not found in --csv-input list")
-    if not os.path.exists(kernel_summary_file):
-        raise FileExistsError(f"{kernel_summary_file} does not exist")
-
-    return pd.read_csv(kernel_summary_file)
+    return pd.read_csv(filename)
